// src/api/models/CreateLevelDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateLevelDto = {
    /**
     * 等级唯一键
     */
    key: string;
    /**
     * 显示名称
     */
    label: string;
    /**
     * 排序/权重
     */
    rank?: number;
    /**
     * 描述
     */
    description?: string;
};