// src/api/models/UpdateLevelDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UpdateLevelDto = {
    /**
     * 显示名称
     */
    label?: string;
    /**
     * 排序/权重
     */
    rank?: number;
    /**
     * 描述
     */
    description?: string;
    /**
     * 是否启用
     */
    isActive?: boolean;
};